requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
idna==2.10
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10